renamed a folder and deleted some unused/irrelevent scripts
