+++ Runtime/ModuleContracts/IApplyDelta.cs.meta
