+++ Runtime/ModuleContracts/IChangeResolver.cs
@@ -3,7 +3,7 @@
 using Spellbound.Core.ObjectData;
 using Spellbound.Core.ObjectHandling;
 
-namespace Spellbound.Core.PresetContracts {
+namespace Spellbound.Core.ModuleContracts {
     /// <summary>
     /// For ObjectPreset PresetModules
     /// Interface Contract for a Module to respond to changes with major consequences
