+++ Runtime/ModuleContracts/IDefaultDataProvider.cs.meta
