+++ Runtime/ModuleContracts/IDeleteHandler.cs
@@ -2,7 +2,7 @@
 
 using Spellbound.Core.ObjectHandling;
 
-namespace Spellbound.Core.PresetContracts {
+namespace Spellbound.Core.ModuleContracts {
     public interface IDeleteHandler {
         /// <summary>
         /// For ObjectPreset PresetModules
