+++ Runtime/ModuleContracts/IDispatch.cs
@@ -4,7 +4,7 @@ using Spellbound.Core.ObjectData;
 using Spellbound.Core.ObjectHandling;
 using Spellbound.Core.Surfaces;
 
-namespace Spellbound.Core.PresetContracts {
+namespace Spellbound.Core.ModuleContracts {
     /// <summary>
     /// For ObjectPreset PresetModules
     /// Interface Contract for a Module to receive a dispatch event from an event surface.
