--- Runtime/Modules/DamageableModule.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c2667fe27a051f049968c646bdaf8f42
