--- Runtime/Modules/InteractableModule.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: dd00a3c154981ba48a4f492ea5828f86
