--- Runtime/Modules/MouseoverModule.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 916820262b2c6b24c9433cb97aa81045
